organisation des routes logout et update_user
